requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
django-filter==23.5
